Revert "fix warnings"
This reverts commit 84723f815eb4aa77727cd5b8e50e3d65a82d10f2.

+++ Lab3/TvSet/TvSet.cpp
@@ -107,7 +107,7 @@ bool CTvSet::SetChannelName(int number, string name)
 	return false;
 }
 
-size_t CTvSet::GetCurrChannel() const
+int CTvSet::GetCurrChannel() const
 {
 	return m_isOn ? m_currChannel : 0;
 }
@@ -123,7 +123,7 @@ bool CTvSet::DeleteChannelName(string name)
 	map<string, size_t>::iterator itName = channelName.find(name);
 	if (m_isOn && itName != channelName.end())
 	{
-		size_t number = itName->second;
+		int number = itName->second;
 		channelName.erase(itName);
 		channelNumber.erase(number);
 		return true;
